chore: handle when current-version does not exist

+++ _includes/includes/page-version.php
@@ -8,6 +8,8 @@
     public $versions, $basePath, $versionedPath, $subPath, $path;
     public $currentVersion;
 
+    public $file, $currentFolder, $currentFileExists;
+
     function __construct() {
       $url = $_SERVER['REQUEST_URI'];
 
@@ -18,9 +20,18 @@
         if(sizeof($this->versions) == 0) {
           $this->active = false;
         } else {
-            $this->currentVersion = $this::getCurrentVersion($this->versions);
-            if($this->versionedPath == 'current-version') $this->versionedPath = $this->currentVersion;
-            if($this->versionedPath == 'latest-version') $this->versionedPath = $this->versions[0];
+          $this->currentVersion = $this::getCurrentVersion($this->versions);
+          if($this->versionedPath == 'current-version') $this->versionedPath = $this->currentVersion;
+          if($this->versionedPath == 'latest-version') $this->versionedPath = $this->versions[0];
+
+          if($this->subPath == '' || $this->subPath[strlen($this->subPath)-1] == '/') {
+            $this->file = $this->subPath.'index';
+          } else {
+            $this->file = $this->subPath;
+          }
+
+          $this->currentFolder = "{$_SERVER['DOCUMENT_ROOT']}{$this->basePath}/{$this->currentVersion}/";
+          $this->currentFileExists = file_exists("{$this->currentFolder}{$this->file}.php") || file_exists("{$this->currentFolder}/{$this->file}.md");
         }
       }
     }
@@ -43,19 +54,8 @@
     ///
     static function WriteHead() {
       global $PageVersion;
-      if(!$PageVersion->active) return;
-
-      if($PageVersion->subPath == '' || $PageVersion->subPath[strlen($PageVersion->subPath)-1] == '/') {
-        $file = $PageVersion->subPath.'index';
-      } else {
-        $file = $PageVersion->subPath;
-      }
-
-      $folder = "{$_SERVER['DOCUMENT_ROOT']}{$PageVersion->basePath}/{$PageVersion->currentVersion}/";
-      if(file_exists("{$folder}{$file}.php") || file_exists("{$folder}/{$file}.md")) {
-        // Only give a canonical link if the content exists in the current version
-        echo "<link rel='canonical' href='{$PageVersion->basePath}/current-version/{$PageVersion->subPath}'>";
-      }
+      if(!$PageVersion->active || !$PageVersion->currentFileExists) return;
+      echo "<link rel='canonical' href='{$PageVersion->basePath}/current-version/{$PageVersion->subPath}'>";
     }
 
     ///
@@ -65,7 +65,7 @@
       global $PageVersion;
       if(!$PageVersion->active) return;
 
-      if($PageVersion->isCurrentPageCurrent()) {
+      if($PageVersion->isCurrentPageCurrent() || !$PageVersion->currentFileExists) {
         // Don't show the banner if we are on the current version
         return;
       }
@@ -97,12 +97,6 @@
       global $PageVersion;
       if(!$PageVersion->active) return;
 
-      if($PageVersion->subPath == '' || $PageVersion->subPath[strlen($PageVersion->subPath)-1] == '/') {
-        $file = $PageVersion->subPath.'index';
-      } else {
-        $file = $PageVersion->subPath;
-      }
-
       // Build the menu
       $html = "<div id='version-selector'>Other versions";
       $html .= "<div>";
@@ -111,7 +105,7 @@
 
       foreach($PageVersion->versions as $version) {
         $folder = "{$_SERVER['DOCUMENT_ROOT']}{$PageVersion->basePath}/{$version}/";
-        if(!file_exists("{$folder}{$file}.php") && !file_exists("{$folder}/{$file}.md")) {
+        if(!file_exists("{$folder}{$PageVersion->file}.php") && !file_exists("{$folder}/{$PageVersion->file}.md")) {
           // Skip links to versions that don't have this exact page
           continue;
         } else {
